backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
open3d==0.19.0
trimesh==4.0.5
redis==5.0.1
requests==2.31.0
numpy==1.24.3
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.0